requirements.txt
Appium-Python-Client==2.11.1
pytest==7.4.3
allure-pytest==2.13.2
selenium==4.11.2
pytest-xdist==3.3.1 